# docs/swagger.yaml
basePath: /
definitions:
  entities.Domain:
    properties:
      domain:
        type: string
      domain_id:
        type: string
      name:
        type: string
    type: object
...
  handlers.AuthResponse:
    properties:
      token:
        type: string
      user:
        properties:
          domain:
            properties:
              description:
                type: string
              id:
                type: string
              name:
                type: string
            type: object
          email:
            type: string
          first_name:
            type: string
          id:
            type: string
          last_name:
            type: string
          role:
            properties:
              claims:
                additionalProperties: true
                type: object
              description:
                type: string
              id:
                type: string
              name:
                type: string
            type: object
          username:
            type: string
        type: object
    type: object
  handlers.CreateDomainRequest:
    properties:
      domain:
        type: string
      name:
        type: string
    required:
    - domain
    - name
    type: object
...
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  handlers.UpdateDomainRequest:
    properties:
      domain:
        type: string
      name:
        type: string
    required:
    - domain
    - name
    type: object
...
  repositories.DomainListResult:
    properties:
      domains:
        items:
          $ref: '#/definitions/entities.Domain'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for Nusarithm IAM Backend
  title: Nusarithm IAM API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Domain ID
        in: header
        name: X-NRM-DID
        required: true
        type: string
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/profile:
    get:
      consumes:
      - application/json
      description: Get authenticated user's profile information
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user profile
      tags:
      - auth
  /auth/validate:
    post:
      consumes:
      - application/json
      description: Validate JWT token and return user information
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Validate JWT token
      tags:
      - auth
  /domains:
    get:
      consumes:
      - application/json
      description: Get all domains with pagination and search
      parameters:
      - description: Search term for domain name
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repositories.DomainListResult'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all domains
      tags:
      - domains
    post:
      consumes:
      - application/json
      description: Create a new domain
      parameters:
      - description: Domain data
        in: body
        name: domain
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDomainRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Domain'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a domain
      tags:
      - domains
  /domains/{domainId}:
    delete:
      consumes:
      - application/json
      description: Delete domain by ID
      parameters:
      - description: Domain ID
        in: path
        name: domainId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a domain
      tags:
      - domains
    get:
      consumes:
      - application/json
      description: Get domain by ID
      parameters:
      - description: Domain ID
        in: path
        name: domainId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Domain'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a domain
      tags:
      - domains
    put:
      consumes:
      - application/json
      description: Update domain by ID
      parameters:
      - description: Domain ID
        in: path
        name: domainId
        required: true
        type: string
      - description: Domain data
        in: body
        name: domain
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Domain'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a domain
      tags:
      - domains
...
swagger: "2.0"